+++ glotaran/builtin/io/__init__.py
@@ -1,0 +1,1 @@
+"""Package containing the builtin IO plugins."""