+++ glotaran/builtin/io/netCDF/__init__.py
@@ -1,0 +1,1 @@
+"""Package containing the NetCDF4 Data IO plugin."""