+++ glotaran/builtin/io/netCDF/netCDF.py
@@ -1,3 +1,4 @@
+"""Module containing the NetCDF4 Data IO plugin."""
 from __future__ import annotations
 
 # Needed to prevent a netCDF4 RuntimeWarning at import time
@@ -11,7 +12,20 @@ from glotaran.io import register_data_io
 
 @register_data_io("nc")
 class NetCDFDataIo(DataIoInterface):
+    """Plugin for NetCDF4 data io."""
+
     def load_dataset(self, file_name: str) -> xr.Dataset | xr.DataArray:
+        """Load a ``*.nc`` file into a :xarraydoc:`Dataset` or :xarraydoc:`DataArray`.
+
+        Parameters
+        ----------
+        file_name: str
+            Path to the ``*.nc`` file that should be loaded.
+
+        Returns
+        -------
+        xr.Dataset | xr.DataArray
+        """
         with xr.open_dataset(file_name) as ds:
             return ds.load()
 
@@ -22,6 +36,16 @@ class NetCDFDataIo(DataIoInterface):
         *,
         data_filters: list[str] | None = None,
     ):
+        """Write a :xarraydoc:`Dataset` to the ``*.nc`` at path ``file_name``.
 
+        Parameters
+        ----------
+        dataset: xr.Dataset
+            :xarraydoc:`Dataset` that should be written to file.
+        file_name: str
+            Path of the file to write ``dataset`` to.
+        data_filters: list[str] | None
+            List of data variable names that should be written to file. Defaults to None.
+        """
         data_to_save = dataset if data_filters is None else dataset[data_filters]
         data_to_save.to_netcdf(file_name, mode="w")