+++ glotaran/builtin/io/yml/__init__.py
@@ -1,0 +1,1 @@
+"""Package containing the YAML Data and Project IO plugins."""